refactor: code-split freehand crop region
